+++ src/App.tsx
@@ -1,15 +1,16 @@
 import { Typography } from "@material-ui/core";
 import React from "react";
 import "./App.css";
+import { LoginPage } from "./app/Login/Login";
 
 function App() {
-  return (
-    <div className="App">
-      <Typography variant="h1" color="textSecondary">
-        INSERT EDUGATOR HERE
-      </Typography>
-    </div>
-  );
+	return (
+		<div className="App">
+			<Typography variant="h1" color="textSecondary">
+				<LoginPage></LoginPage>
+			</Typography>
+		</div>
+	);
 }
 
 export default App;
