+++ src/app/common/rootSaga.ts
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import authSaga from "../auth/auth.sagas";
+import authSaga from "../Login/auth.sagas";
 
 export default function* rootSaga() {
 	yield all([authSaga()]);
